commit 33: restructured contracts dto:one file one class
