+++ Polyclinic/Polyclinic.Contracts/Dto/CountResponseDto.cs
@@ -1,0 +1,12 @@
+namespace Polyclinic.Contracts.Dto;
+
+/// <summary>
+/// Count of Dto entities  
+/// </summary>
+public class CountResponseDto
+{
+    /// <summary>
+    /// Dto count  
+    /// </summary>
+    public int Count { get; set; }
+}
